requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
mysql-connector-python==8.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
python-dateutil==2.8.2